chore: move sortByKey to helpers
