+++ src/helpers/sortByKey.js
@@ -1,0 +1,11 @@
+const sortByKey = (array, key) => {
+  // Return a copy of array of objects sorted
+  // by the given key
+  return array.sort(function (a, b) {
+    var x = a[key]
+    var y = b[key]
+    return x < y ? -1 : x > y ? 1 : 0
+  })
+}
+
+export default sortByKey
