+++ src/utils.js
@@ -3,16 +3,6 @@ import { isEmpty, isEqual, xorWith, cloneDeep, sortBy } from 'lodash'
 const arrayEquals = (x, y) => isEmpty(xorWith(x, y, isEqual))
 const deepCopy = (obj) => cloneDeep(obj)
 
-const sortByKey = (array, key) => {
-  // Return a copy of array of objects sorted
-  // by the given key
-  return array.sort(function (a, b) {
-    var x = a[key]
-    var y = b[key]
-    return x < y ? -1 : x > y ? 1 : 0
-  })
-}
-
 const isGroupable = (data, key, value) => {
   // Returns true if the key with the given value is
   // presented multiple times in the array
@@ -179,7 +169,6 @@ export {
   arrayEquals,
   deepCopy,
   groupResult,
-  sortByKey,
   getFolderTypeIcon,
   getTaskTypeIcon,
   getStatusProps,
